Move equals method to separate module
