+++ js/integration/angular/component_registrator.js
@@ -16,6 +16,7 @@ import { NgTemplate } from "./template";
 import ngModule from "./module";
 import CollectionWidget from "../../ui/collection/ui.collection_widget.edit";
 import dataUtils from "../../core/utils/data";
+import { equals } from "../../core/utils/comparator";
 const compileSetter = dataUtils.compileSetter;
 const compileGetter = dataUtils.compileGetter;
 import extendUtils from "../../core/utils/extend";
@@ -163,7 +164,7 @@ let ComponentBuilder = Class.inherit({
                 this._component.option(optionPath, newValue);
                 updateWatcher();
 
-                if(dataUtils.equals(oldValue, newValue) && this._ngLocker.locked(optionPath)) {
+                if(equals(oldValue, newValue) && this._ngLocker.locked(optionPath)) {
                     this._ngLocker.release(optionPath);
                 }
             };
