+++ js/ui/grid_core/ui.grid_core.adaptivity.js
@@ -10,7 +10,7 @@ import Form from "../form";
 import gridCoreUtils from "./ui.grid_core.utils";
 import themes from "../themes";
 import { getWindow } from "../../core/utils/window";
-import { equalByComplexValue } from "../../core/utils/data";
+import { equalByValue } from "../../core/utils/common";
 import { each } from "../../core/utils/iterator";
 import { extend } from "../../core/utils/extend";
 import { Deferred, when } from "../../core/utils/deferred";
@@ -315,7 +315,7 @@ var AdaptiveColumnsController = modules.ViewController.inherit({
 
         if(this._isRowEditMode()) {
             var editRowKey = this._editingController.getEditRowKey();
-            if(equalByComplexValue(editRowKey, this._dataController.adaptiveExpandedKey())) {
+            if(equalByValue(editRowKey, this._dataController.adaptiveExpandedKey())) {
                 return true;
             }
         } else {
@@ -662,7 +662,7 @@ var AdaptiveColumnsController = modules.ViewController.inherit({
      * @return boolean
      */
     isAdaptiveDetailRowExpanded: function(key) {
-        return this._dataController.adaptiveExpandedKey() && equalByComplexValue(this._dataController.adaptiveExpandedKey(), key);
+        return this._dataController.adaptiveExpandedKey() && equalByValue(this._dataController.adaptiveExpandedKey(), key);
     },
 
     /**
