+++ js/ui/grid_core/ui.grid_core.validating.js
@@ -2,8 +2,7 @@ import $ from "../../core/renderer";
 import eventsEngine from "../../events/core/events_engine";
 import modules from "./ui.grid_core.modules";
 import { createObjectWithChanges, getIndexByKey } from "./ui.grid_core.utils";
-import { grep, deferUpdate } from "../../core/utils/common";
-import { equalByComplexValue } from "../../core/utils/data";
+import { grep, deferUpdate, equalByValue } from "../../core/utils/common";
 import { each } from "../../core/utils/iterator";
 import { isDefined } from "../../core/utils/type";
 import { extend } from "../../core/utils/extend";
@@ -370,7 +369,7 @@ module.exports = {
                                 key = editData.key;
 
                             each(items, function(i, item) {
-                                if(equalByComplexValue(key, isInsert ? item : dataController.keyOf(item))) {
+                                if(equalByValue(key, isInsert ? item : dataController.keyOf(item))) {
                                     index = i;
                                     return false;
                                 }
