+++ js/ui/tag_box.js
@@ -4,8 +4,7 @@ import dataUtils from "../core/element_data";
 import eventsEngine from "../events/core/events_engine";
 import registerComponent from "../core/component_registrator";
 import browser from "../core/utils/browser";
-import { noop, ensureDefined } from "../core/utils/common";
-import { equalByComplexValue } from "../core/utils/data";
+import { noop, ensureDefined, equalByValue } from "../core/utils/common";
 import { SelectionFilterCreator as FilterCreator } from "../core/utils/selection_filter";
 import { Deferred, when } from "../core/utils/deferred";
 import { getPublicElement, createTextElementHiddenCopy } from "../core/utils/dom";
@@ -1081,7 +1080,7 @@ const TagBox = SelectBox.inherit({
             const $tag = $tags[i],
                 tagData = dataUtils.data($tag, TAGBOX_TAG_DATA_KEY);
 
-            if(value === tagData || (equalByComplexValue(value, tagData))) {
+            if(value === tagData || (equalByValue(value, tagData))) {
                 result = $($tag);
                 break;
             }
